src/test/javascript/spec/app/entities/issued-stock-master/issued-stock-master.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import { IssuedStockMasterService } from 'app/entities/issued-stock-master/issued-stock-master.service';
import { IIssuedStockMaster, IssuedStockMaster } from 'app/shared/model/issued-stock-master.model';

describe('Service Tests', () => {
  describe('IssuedStockMaster Service', () => {
    let injector: TestBed;
    let service: IssuedStockMasterService;
    let httpMock: HttpTestingController;
    let elemDefault: IIssuedStockMaster;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(IssuedStockMasterService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new IssuedStockMaster(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a IssuedStockMaster', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new IssuedStockMaster(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a IssuedStockMaster', () => {
        const returnedFromService = Object.assign(
          {
            flockNumber: 'BBBBBB',
            materialCode: 'BBBBBB',
            materialName: 'BBBBBB',
            stockIssued: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of IssuedStockMaster', () => {
        const returnedFromService = Object.assign(
          {
            flockNumber: 'BBBBBB',
            materialCode: 'BBBBBB',
            materialName: 'BBBBBB',
            stockIssued: 1
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a IssuedStockMaster', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
